requirements.txt
asgiref==3.5.2
Django==4.0.5
django-debug-toolbar==3.5.0
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
